make readable available in ChannelBase
Channel<T> is now derived from ChannelBase, which can be used to wait
for a channel without knowing its type (which will be useful to reduce
the code size of Actor::wait etc.

ChannelImpl is also split up into ChannelStoreBase and ChannelStore<T>
to facilitate this (renamed to store to reflect that it's actual role)

This used a custom queue implementation embedded into ChannelStore<T>
to allow access to the size from the untyped base, which isn't possible
with std::queue without RTTI (and is very awkward with any external
class).

This perhaps moves away from the idea of not doing anything clever and
building on top of basic C++ stuff, but I think it still meets the
underlying goal (being portable and correct). We'll see.

This should be re-arranged a bit to not bury the interface below a load
of implementation.